backend/pyproject.toml
[project]
name = "moovup-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "email-validator>=2.1",
    "sqlalchemy>=2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "sentence-transformers>=2.7",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "pandas>=2.2",
    "beautifulsoup4>=4.12",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src", "."]
